+++ include/gui_fms_status.h
@@ -18,7 +18,7 @@ public:
 private:
     static void create(void);
     static void hotkey_handler(void * inRefcon);
-    static int	click(XPWidgetMessage inMessage,XPWidgetID inWidget,long inParam1,long inParam2);
+    static int  click(XPWidgetMessage inMessage,XPWidgetID inWidget,intptr_t inParam1,intptr_t inParam2);
 };
 
 #endif // GUI_FMS_STATUS_H
